refactor(tui::ui::components::popups): move save-playlist related popups to own file
