+++ tui/src/ui/components/popups/mod.rs
@@ -38,6 +38,7 @@ use tuirealm::{AttrValue, Attribute, Component, Event, MockComponent, State, Sta
 mod error;
 mod help;
 mod podcast;
+mod saveplaylist;
 
 pub use error::ErrorPopup;
 pub use help::HelpPopup;
@@ -46,6 +47,7 @@ pub use podcast::{
     FeedDeleteConfirmInputPopup, FeedDeleteConfirmRadioPopup, PodcastAddPopup,
     PodcastSearchTablePopup,
 };
+pub use saveplaylist::{SavePlaylistConfirm, SavePlaylistPopup};
 
 #[derive(MockComponent)]
 pub struct QuitPopup {
@@ -373,195 +375,6 @@ impl Component<Msg, NoUserEvent> for MessagePopup {
         //     _ => None,
         // }
         None
-    }
-}
-
-#[derive(MockComponent)]
-pub struct SavePlaylistPopup {
-    component: Input,
-}
-
-impl SavePlaylistPopup {
-    pub fn new(style_color_symbol: &StyleColorSymbol) -> Self {
-        Self {
-            component: Input::default()
-                .foreground(
-                    style_color_symbol
-                        .library_foreground()
-                        .unwrap_or(Color::Yellow),
-                )
-                .background(
-                    style_color_symbol
-                        .library_background()
-                        .unwrap_or(Color::Reset),
-                )
-                .borders(
-                    Borders::default()
-                        .color(style_color_symbol.library_border().unwrap_or(Color::Green))
-                        .modifiers(BorderType::Rounded),
-                )
-                // .invalid_style(Style::default().fg(Color::Red))
-                .input_type(InputType::Text)
-                .title("Save Playlist as: (Enter to confirm)", Alignment::Left),
-        }
-    }
-}
-
-impl Component<Msg, NoUserEvent> for SavePlaylistPopup {
-    fn on(&mut self, ev: Event<NoUserEvent>) -> Option<Msg> {
-        let cmd_result = match ev {
-            Event::Keyboard(KeyEvent {
-                code: Key::Left, ..
-            }) => self.perform(Cmd::Move(Direction::Left)),
-            Event::Keyboard(KeyEvent {
-                code: Key::Right, ..
-            }) => self.perform(Cmd::Move(Direction::Right)),
-            Event::Keyboard(KeyEvent {
-                code: Key::Home, ..
-            }) => self.perform(Cmd::GoTo(Position::Begin)),
-            Event::Keyboard(KeyEvent { code: Key::End, .. }) => {
-                self.perform(Cmd::GoTo(Position::End))
-            }
-            Event::Keyboard(KeyEvent {
-                code: Key::Delete, ..
-            }) => self.perform(Cmd::Cancel),
-            Event::Keyboard(KeyEvent {
-                code: Key::Backspace,
-                ..
-            }) => {
-                self.perform(Cmd::Delete);
-                self.perform(Cmd::Submit)
-            }
-            Event::Keyboard(KeyEvent {
-                code: Key::Char(ch),
-                modifiers: KeyModifiers::SHIFT | KeyModifiers::NONE,
-            }) => {
-                self.perform(Cmd::Type(ch));
-                self.perform(Cmd::Submit)
-            }
-            Event::Keyboard(KeyEvent { code: Key::Esc, .. }) => {
-                return Some(Msg::SavePlaylistPopupCloseCancel);
-            }
-            Event::Keyboard(KeyEvent {
-                code: Key::Enter, ..
-            }) => match self.component.state() {
-                State::One(StateValue::String(input_string)) => {
-                    return Some(Msg::SavePlaylistPopupCloseOk(input_string))
-                }
-                _ => return Some(Msg::None),
-            },
-            _ => CmdResult::None,
-        };
-        match cmd_result {
-            CmdResult::Submit(State::One(StateValue::String(input_string))) => {
-                Some(Msg::SavePlaylistPopupUpdate(input_string))
-            }
-            _ => Some(Msg::None),
-        }
-    }
-}
-
-#[derive(MockComponent)]
-pub struct SavePlaylistConfirm {
-    component: Radio,
-    config: SharedSettings,
-    filename: String,
-}
-
-impl SavePlaylistConfirm {
-    pub fn new(config: SharedSettings, filename: &str) -> Self {
-        let component = {
-            let config = config.read();
-            Radio::default()
-                .foreground(
-                    config
-                        .style_color_symbol
-                        .library_foreground()
-                        .unwrap_or(Color::Yellow),
-                )
-                .background(
-                    config
-                        .style_color_symbol
-                        .library_background()
-                        .unwrap_or(Color::Reset),
-                )
-                .borders(
-                    Borders::default()
-                        .color(
-                            config
-                                .style_color_symbol
-                                .library_border()
-                                .unwrap_or(Color::Yellow),
-                        )
-                        .modifiers(BorderType::Rounded),
-                )
-                .title(
-                    // format!(" Playlist {filename} exists. Overwrite? "),
-                    " Playlist exists. Overwrite? ",
-                    Alignment::Center,
-                )
-                .rewind(true)
-                .choices(&["No", "Yes"])
-                .value(0)
-        };
-
-        Self {
-            component,
-            config,
-            filename: filename.to_string(),
-        }
-    }
-}
-
-impl Component<Msg, NoUserEvent> for SavePlaylistConfirm {
-    fn on(&mut self, ev: Event<NoUserEvent>) -> Option<Msg> {
-        let config = self.config.clone();
-        let keys = &config.read().keys;
-        let cmd_result = match ev {
-            Event::Keyboard(KeyEvent {
-                code: Key::Left, ..
-            }) => self.perform(Cmd::Move(Direction::Left)),
-            Event::Keyboard(KeyEvent {
-                code: Key::Right, ..
-            }) => self.perform(Cmd::Move(Direction::Right)),
-
-            Event::Keyboard(key) if key == keys.global_left.key_event() => {
-                self.perform(Cmd::Move(Direction::Left))
-            }
-            Event::Keyboard(key) if key == keys.global_right.key_event() => {
-                self.perform(Cmd::Move(Direction::Right))
-            }
-            Event::Keyboard(key) if key == keys.global_up.key_event() => {
-                self.perform(Cmd::Move(Direction::Left))
-            }
-            Event::Keyboard(key) if key == keys.global_down.key_event() => {
-                self.perform(Cmd::Move(Direction::Right))
-            }
-            Event::Keyboard(key) if key == keys.global_quit.key_event() => {
-                return Some(Msg::SavePlaylistConfirmCloseCancel)
-            }
-            Event::Keyboard(key) if key == keys.global_esc.key_event() => {
-                return Some(Msg::SavePlaylistConfirmCloseCancel)
-            }
-
-            Event::Keyboard(KeyEvent {
-                code: Key::Enter, ..
-            }) => self.perform(Cmd::Submit),
-            _ => return None,
-        };
-        if matches!(
-            cmd_result,
-            CmdResult::Submit(State::One(StateValue::Usize(0)))
-        ) {
-            Some(Msg::SavePlaylistConfirmCloseCancel)
-        } else if matches!(
-            cmd_result,
-            CmdResult::Submit(State::One(StateValue::Usize(1)))
-        ) {
-            Some(Msg::SavePlaylistConfirmCloseOk(self.filename.clone()))
-        } else {
-            Some(Msg::None)
-        }
     }
 }
